# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Building Charges & Payments API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all user accounts
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedUser"
        "403":
          description: Forbidden (admin only)
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user account (email + temporary password + role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "400":
          description: Validation error
        "403":
          description: Forbidden (admin only)
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user's role or disabled state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "400":
          description: Validation error
        "403":
          description: Forbidden (admin only)
        "404":
          description: Not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
        "400":
          description: Cannot delete own account
        "403":
          description: Forbidden (admin only)
        "404":
          description: Not found
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset a user's password to a new temporary password (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserResetPasswordInput"
      responses:
        "200":
          description: Password reset successfully
        "400":
          description: Validation error
        "403":
          description: Forbidden (admin only)
        "404":
          description: Not found
...
components:
  schemas:
...
    ManagedUser:
      type: object
      required: [id, email, role, disabled]
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [admin, accountant, viewer]
        disabled:
          type: boolean
        createdAt:
          type: string
        passwordChangedAt:
          type: string
          description: >-
            Effective last password-change timestamp from Supabase
            user_metadata.password_changed_at, falling back to the account
            creation time. Absent if the Supabase account could not be resolved.
    UserCreateInput:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, accountant, viewer]
    UserUpdateInput:
      type: object
      properties:
        role:
          type: string
          enum: [admin, accountant, viewer]
        disabled:
          type: boolean
    UserResetPasswordInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 6
...