# lib/api-zod/src/generated/types/userResetPasswordInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Building Charges & Payments API
 * OpenAPI spec version: 0.1.0
 */

export interface UserResetPasswordInput {
  /** @minLength 6 */
  password: string;
}